+++ CHANGELOG.md
@@ -2,3 +2,4 @@ __0.3.0__
 
 *   Rename `SearchProcessorInterface::search` to `constraintBuilder` to support Laravel 4.2
 *   Support Laravel 4.2
+*   Add `states` column type for defining states for rows